src/functions/requirements.txt
azure-functions==1.18.0
anthropic==0.39.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-servicebus==7.11.4
azure-identity==1.16.1
pillow==10.3.0
python-dotenv==1.0.0
pydantic==2.6.1
aiohttp==3.10.11
